Add a2a3sim platform for host-based simulation
Add thread-based simulation platform (a2a3sim) that emulates AICPU/AICore
execution on host, enabling development and testing without Ascend hardware.

Platform changes:
- Add a2a3sim platform with host/aicore/aicpu/common modules
- Implement device_runner, memory_allocator, function_cache, and C API
- Use snake_case naming convention throughout the platform

Tooling and build:
- Update binary_compiler and pto_compiler to support multi-platform builds
- Add ASCEND_HOME_PATH environment variable validation in binary_compiler
- Add toolchain module for platform-aware compilation
- Rename runtime_bindings.py to bindings.py for clarity

Examples and docs:
- Add host_build_graph_sim_example demonstrating simulation usage
- Update README with platform comparison table
- Update host_build_graph_example for API consistency

+++ src/runtime/host_build_graph/aicpu/aicpu_executor.cpp
@@ -187,7 +187,7 @@ int AicpuExecutor::hank_aicore(Runtime* runtime, int thread_idx, const int* cur_
         Handshake* hank = &all_hanks[core_id];
         while (hank->aicore_done == 0) {
         }
-        DEV_INFO("Thread %d: success hank->aicore_done = %u", thread_idx, (uint64_t)hank->aicore_done);
+        DEV_INFO("Thread %d: success hank->aicore_done = %u", thread_idx, hank->aicore_done);
     }
     return 0;
 }
